// src/type.d.ts
import { type } from "os"

interface IArticle {
    id: number,
    title: string, 
    body: string,
    image: string,
    imageUrl: string,
    imageName: string
}

type ArticleState = {
    articles : IArticle[]
}

type ArticleAction = {
    type: string,
    article: IArticle
}

type DispatchType = (args: ArticleAction) => ArticleAction